Animation and Physics adjustments
Added a simple animation to the compass to make it wobble, and looked into the physics of the boat and tightened it up to make it work more the way I want it to. Attempted to add an animation to tilt the ship on a turn, but it needs more work.