Added Main Menu
Main menu now works with an options menu hooked up. I'm particularly fond of the background music persisting between the menus.

Future steps: Make the Main Menu the first thing when booting up the game, after logos and stuff of course.

Add the ship as a background element, floating on the waves.